pyproject.toml
[tool.poetry]
name = "ot-harjoitustyo"
version = "0.1.0"
description = ""
authors = ["Timo Paloneva <timo.paloneva@gmail.com>"]
readme = "README.md"
packages = [{include = "ot_harjoitustyo"}]

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^1.0.0"
invoke = "^2.0.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"
coverage = "^7.2.2"
pylint = "^2.17.2"
autopep8 = "^2.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
